pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-grocery-price-scanner"
version = "0.1.0"
description = "A multi-agent system for scraping grocery prices and optimizing shopping lists"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=0.1.0",
    "qdrant-client>=1.7.0",
    "ollama>=0.1.7",
    "streamlit>=1.28.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
grocery-scanner = "agentic_grocery_price_scanner.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/agentic-grocery-price-scanner"
Repository = "https://github.com/yourusername/agentic-grocery-price-scanner.git"
Issues = "https://github.com/yourusername/agentic-grocery-price-scanner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_grocery_price_scanner*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=agentic_grocery_price_scanner --cov-report=term-missing"